requirements.txt
langchain-community
faiss-gpu
sentence_transformers 
fastapi
autocorrect
nltk
sqlalchemy
bitsandbytes